requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
reportlab==4.0.7
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0